requirements.txt
pybit
okx-sdk
python-telegram-bot
python-dotenv
apscheduler
websockets
pytz
t-tech-investments --index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple
